Upgrade to apollo-server-2
This removes koa, but we can add it back should we need it again
Apolo-server-2 comes with playground, which is better than graphiql
