+++ api/data/package.json
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "ci": "yarn lint && yarn test",
-    "start": "cross-env NODE_ENV=development nodemon src --exec babel-node --watch src | bunyan -o short --color",
+    "start": "cross-env NODE_ENV=development babel-node src | bunyan -o short --color",
     "test": "jest --config jest.config.js --coverage",
     "test:watch": "jest --config jest.config.js --watch",
     "lint": "eslint . --ignore-path ../../.gitignore",
@@ -27,25 +27,18 @@
   ],
   "private": true,
   "dependencies": {
-    "apollo-server-koa": "^1.3.2",
+    "apollo-server": "^2.0.5",
     "axios": "^0.18.0",
     "bluebird": "^3.5.1",
-    "boom": "^7.2.0",
     "bunyan": "^1.8.12",
     "cross-env": "^5.1.3",
     "dotenv": "^5.0.1",
     "fs-extra": "^5.0.0",
-    "graphql": "^0.13.1",
-    "graphql-playground-middleware-koa": "^1.4.2",
+    "graphql": "^0.13.2",
     "graphql-tools": "^2.21.0",
     "knex": "^0.14.4",
-    "koa": "^2.5.0",
-    "koa-bodyparser": "^4.2.0",
     "koa-bunyan-logger": "^2.0.0",
-    "koa-compose": "^4.0.0",
-    "koa-router": "^7.4.0",
     "lodash": "^4.17.5",
-    "ramda": "^0.25.0",
     "sanitize-filename": "^1.6.1",
     "sqlite3": "^3.1.13"
   },
@@ -65,8 +58,6 @@
     "eslint-plugin-import": "^2.9.0",
     "eslint-plugin-prettier": "^2.6.0",
     "jest": "^22.4.2",
-    "nock": "^9.2.3",
-    "nodemon": "^1.17.1",
     "prettier": "^1.11.1"
   }
 }
