+++ api/data/src/index.js
@@ -1,41 +1,111 @@
-import Koa from 'koa';
-
-import Router from 'koa-router';
-import bodyParser from 'koa-bodyparser';
-import { graphqlKoa, graphiqlKoa } from 'apollo-server-koa';
-import koaPlayground from 'graphql-playground-middleware-koa';
-
-import Boom from 'boom';
-import loggerMiddleware from 'koa-bunyan-logger';
-import errorMiddleware from './middleware/error';
-
+import { ApolloServer, gql } from 'apollo-server';
+import _ from 'lodash';
+import playgroundConfig from './config/graphqlPlayground';
 import log from './util/log';
-import schema from './graphql';
+import jsonData from './jsonData';
 
-const app = new Koa();
-const router = new Router();
+const typeDefs = gql`
+  # Describes a module for, may span different semesters
+  type Module {
+    code: String!
+    title: String!
+    department: String
+    description: String
+    credit: Float
+    workload: String
+    prerequisite: String
+    corequisite: String
+    corsBiddingStats: [CorsBiddingStats]
+    # Refers to the history of the module throughout semesters
+    history: [ModuleInfo]!
+  }
 
-// Register middleware
-app.use(bodyParser());
-app.use(loggerMiddleware(log));
-app.use(loggerMiddleware.requestIdContext());
-app.use(loggerMiddleware.requestLogger());
-app.use(errorMiddleware());
+  # Describes a particular module for a semester
+  type ModuleInfo {
+    semester: Int
+    examDate: String
+    examOpenBook: Boolean
+    examDuration: String
+    examVenue: String
+    timetable: [Lesson]
+  }
 
-// Registers routes
-router.post('/graphql', graphqlKoa({ schema, tracing: true }));
-router.get('/graphiql', graphiqlKoa({ endpointURL: '/graphql' }));
-router.get('/playground', koaPlayground({ endpoint: '/graphql' }));
+  # Bidding stats for Cors
+  type CorsBiddingStats {
+    quota: Int
+    bidders: Int
+    lowestBid: Int
+    lowestSuccessfulBid: Int
+    highestBid: Int
+    faculty: String
+    studentAcctType: String
+    acadYear: String
+    semester: Int
+    round: String
+    group: String
+  }
 
-app.use(router.routes());
-app.use(
-  router.allowedMethods({
-    throw: true,
-    notImplemented: () => new Boom.notImplemented(), // eslint-disable-line new-cap
-    methodNotAllowed: () => new Boom.methodNotAllowed(), // eslint-disable-line new-cap
-  }),
-);
+  # A lesson conducted, may it be a lecture, laboratory or lecture
+  type Lesson {
+    classNo: String!
+    lessonType: String!
+    weekText: String!
+    dayText: String!
+    startTime: String!
+    endTime: String!
+    venue: String!
+  }
 
-log.info('current environment: %s', process.env.NODE_ENV);
-log.info('server started at port: %d', process.env.PORT || 3600);
-app.listen(process.env.PORT || 3600);
+  # the schema allows the following query:
+  type Query {
+    modules(acadYear: String!, first: Int, offset: Int): [Module!]!
+    module(acadYear: String!, code: String!): Module
+  }
+
+  schema {
+    query: Query
+  }
+`;
+
+const resolvers = {
+  Query: {
+    modules(root, { acadYear, first, offset }) {
+      const yearData = jsonData[acadYear];
+      if (yearData == null) {
+        return [];
+      }
+      const modules = Object.values(yearData);
+      return modules.slice(offset, offset ? offset + first : first);
+    },
+    module(root, { acadYear, code }) {
+      return _.get(jsonData, [acadYear, code]);
+    },
+  },
+};
+
+const server = new ApolloServer({
+  typeDefs,
+  /* Apollo is mutating resolvers */
+  resolvers: { ...resolvers },
+  playground: playgroundConfig,
+  formatError: (error) => {
+    log.error(error);
+    return error;
+  },
+  formatResponse: (response) => {
+    log.info(response);
+    return response;
+  },
+});
+
+if (process.env.NODE_ENV !== 'test') {
+  server.listen().then(({ url }) => {
+    log.info(`🚀  Server ready at ${url}`);
+  });
+}
+
+/* For testing purposes */
+export default {
+  typeDefs,
+  resolvers,
+};
